requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
base58==2.1.1
bitarray==1.2.2
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cssselect2==0.4.1
cytoolz==0.11.2
defusedxml==0.7.1
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
frozenlist==1.2.0
hexbytes==0.2.2
idna==3.3
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lru-dict==1.1.7
lxml==4.7.1
multiaddr==0.0.9
multidict==5.2.0
netaddr==0.8.0
parsimonious==0.8.1
Pillow==8.4.0
protobuf==3.19.1
pycparser==2.21
pycryptodome==3.12.0
pyrsistent==0.18.0
python-dotenv==0.19.2
reportlab==3.6.3
requests==2.26.0
rlp==2.0.1
six==1.16.0
svglib==1.1.0
tinycss2==1.1.1
toolz==0.11.2
typing_extensions==4.0.1
urllib3==1.26.7
varint==1.0.2
web3==5.25.0
webencodings==0.5.1
websockets==9.1
yarl==1.7.2
